Fixed Email Formatting

+++ FileMover/FileMover/TVShows.cs
@@ -1,7 +1,4 @@
-﻿using MyEmails;
-using System.Reflection.Metadata.Ecma335;
-using System.Text;
-using System.Text.RegularExpressions;
+﻿using System.Text.RegularExpressions;
 
 namespace FileMover
 {
@@ -9,42 +6,40 @@ namespace FileMover
     {
         public string MoveFile(string fullFilePath, string fileName, string episode)
         {
-            string returnString = string.Empty;
-            string name = string.Empty;
-            string season = string.Empty;
-            
-            if (Regex.Match(fileName, Base.patternSeason1).Success)
+            try
             {
-                season = episode.Substring(1, 2);
-            }
-            else 
-            {
-                season = $"0{episode.Substring(1, 1)}";
-            }
-            name = fileName.Remove(fileName.IndexOf(episode)).Replace(".", " ").TrimEnd();
+                string name = string.Empty;
+                string season = string.Empty;
 
-            var tvShow = Base.data[PropertiesEnum.TV.ToString()];
-            var tvFile = Path.GetFileName(fullFilePath);
-            //foreach (var tvShow in tvShows)
-            //{
+                if (Regex.Match(fileName, Base.patternSeason1).Success)
+                {
+                    season = episode.Substring(1, 2);
+                }
+                else
+                {
+                    season = $"0{episode.Substring(1, 1)}";
+                }
+                name = fileName.Remove(fileName.IndexOf(episode)).Replace(".", " ").TrimEnd();
+
+                var tvShow = Base.data[PropertiesEnum.TV.ToString()];
+                var tvFile = Path.GetFileName(fullFilePath);
+
                 Console.WriteLine($"Starting Copy {tvFile}");
-                try
+
+                if (!File.Exists(@$"{Base.data[PropertiesEnum.TV.ToString()]}{name}\Season {season}\{tvFile}"))
                 {
                     File.Copy(fullFilePath, @$"{Base.data[PropertiesEnum.TV.ToString()]}{name}\Season {season}\{tvFile}");
                     Console.WriteLine($"Finished Copy {tvFile}");
-                    return @$"Copied to {Base.data[PropertiesEnum.TV.ToString()]}{name}\Season {season}\{tvFile}";
+                    return @$"{Base.data[PropertiesEnum.TV.ToString()]}{name}\Season {season}\{tvFile}";
                 }
+                return string.Empty;
+            }
 
-                catch
-                {
-                    File.AppendAllText(Base.data[PropertiesEnum.LogFile.ToString()],
-                        $"Copy Failed {Base.data[PropertiesEnum.TV.ToString()]}{name}\\Season {season}\\{tvFile}\r\n");
-
-                    Console.WriteLine($"Copy Failed {Base.data[PropertiesEnum.TV.ToString()]}{name}\\Season {season}\\{tvFile}");
-                   
-                }
-            //}
-            return returnString;
+            catch (Exception ex)
+            {
+                Console.WriteLine($"Copy Failed {fileName}\r\n{ex}");
+                return $"Copy Failed {fileName}\r\n{ex}";
+            }
         }
     }
 }
